Move MediaStreamTrack into separate module
