+++ crates/native/src/api/media_stream_track/track_state.rs
@@ -1,0 +1,33 @@
+//! Current state of the [`MediaStreamTrack`].
+
+use libwebrtc_sys as sys;
+
+#[cfg(doc)]
+use crate::api::MediaStreamTrack;
+
+/// Indicator of the current [MediaStreamTrackState][0] of a
+/// [`MediaStreamTrack`].
+///
+/// [0]: https://w3.org/TR/mediacapture-streams#dom-mediastreamtrackstate
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum TrackState {
+    /// [MediaStreamTrackState.live][0] representation.
+    ///
+    /// [0]: https://tinyurl.com/w3mcs#idl-def-MediaStreamTrackState.live
+    Live,
+
+    /// [MediaStreamTrackState.ended][0] representation.
+    ///
+    /// [0]: https://tinyurl.com/w3mcs#idl-def-MediaStreamTrackState.ended
+    Ended,
+}
+
+impl From<sys::TrackState> for TrackState {
+    fn from(state: sys::TrackState) -> Self {
+        match state {
+            sys::TrackState::kLive => Self::Live,
+            sys::TrackState::kEnded => Self::Ended,
+            _ => unreachable!(),
+        }
+    }
+}
