+++ lib/src/api/bridge/frb.dart
@@ -5,6 +5,10 @@ export 'api/media_display_info.dart';
 export 'api/media_stream_constraints.dart';
 export 'api/media_stream_constraints/audio_constraints.dart';
 export 'api/media_stream_constraints/video_constraints.dart';
+export 'api/media_stream_track.dart';
+export 'api/media_stream_track/media_type.dart';
+export 'api/media_stream_track/track_event.dart';
+export 'api/media_stream_track/track_state.dart';
 export 'api/rtc_ice_candidate_stats.dart';
 export 'api/rtc_session_description.dart';
 export 'api/rtc_stats.dart';
